requirements.txt
asgiref==3.6.0
astroid==2.15.0
autopep8==2.0.2
colorama==0.4.6
dill==0.3.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
flake8==6.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.1
mccabe==0.7.0
mypy==1.0.1
mypy-extensions==1.0.0
platformdirs==3.1.0
pycodestyle==2.10.0
pyflakes==3.0.1
pylint==2.16.3
pytz==2022.7.1
sqlparse==0.4.3
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
wrapt==1.15.0
